refactor: 클래스 세분화, 정규식 사용, 함수 이름 변경, 단위 테스트 로직 수정
