+++ README.md
@@ -7,7 +7,7 @@
 - [x] 구분자를 기준으로 값이 없는 경우 0으로 인식한다.
 - [x] 문자열 계산기에 숫자 이외의 값 또는 음수를 전달하는 경우 RuntimeException 예외를 throw 한다.
 - [x] 숫자 리스트가 오는 경우 합을 구한다.
-- [x] 공백의 경과 0의 경우 결과값도 0
+- [x] 공백 값을 입력하면 결과값은 0
 
 ## 프로그램 제약 사항
