+++ src/test/kotlin/stringCalculator/StringCalculatorTest.kt
@@ -10,7 +10,7 @@ class StringCalculatorTest : AnnotationSpec() {
     fun `쉽표 또는 콜론을 구분자로 인식한다`() {
         val strings = listOf("1,2,3", "1,2:3", "1:2:3")
         strings.forEach {
-            StringPlusCalculator.seperate(it) shouldBe listOf(1, 2, 3)
+            StringSeperator.seperate(it) shouldBe listOf(1, 2, 3)
         }
     }
 
@@ -18,26 +18,26 @@ class StringCalculatorTest : AnnotationSpec() {
     fun `앞의 기본 구분자(쉼표, 콜론) 외에 커스텀 구분자를 지정할 수 있다`() {
         val strings = listOf("//;\n1;2;3", "1,2:3", "1:2:3", "//;\n1,2;3", "//;\n1,2:3")
         strings.forEach {
-            StringPlusCalculator.seperate(it) shouldBe listOf(1, 2, 3)
+            StringSeperator.seperate(it) shouldBe listOf(1, 2, 3)
         }
     }
 
     @Test
     fun `구분자를 기준으로 값이 없는 경우 0으로 인식`() {
         val string = ""
-        StringPlusCalculator.seperate(string) shouldBe listOf(0)
+        StringSeperator.seperate(string) shouldBe listOf(0)
     }
 
     @Test
     fun `문자열 계산기에 숫자 이외의 값 또는 음수를 전달하는 경우 RuntimeException 예외를 throw`() {
         shouldThrow<RuntimeException> {
             val string = "-1,2:3"
-            StringPlusCalculator.seperate(string)
+            StringSeperator.seperate(string)
         }
 
         shouldThrow<RuntimeException> {
             val string = "a,2:3"
-            StringPlusCalculator.seperate(string)
+            StringSeperator.seperate(string)
         }
     }
 
@@ -48,8 +48,7 @@ class StringCalculatorTest : AnnotationSpec() {
     }
 
     @Test
-    fun `공백의 경과 0의 경우 결과값도 0`() {
-        val data = listOf(0)
-        StringPlusCalculator.calculate(data) shouldBe 0
+    fun `공백 값을 입력하면 결과값은 0`() {
+        StringPlusCalculator.calculate(StringSeperator.seperate("")) shouldBe 0
     }
 }
